Improve architecture more componentized
